+++ src/Exception/NotPositiveException.php
@@ -14,25 +14,19 @@
  */
 namespace Tools\Exception;
 
-use Exception;
+use Tools\Exception\InvalidValueException;
 
 /**
  * "Not positive value" exception
  */
-class NotPositiveException extends Exception
+class NotPositiveException extends InvalidValueException
 {
-    /**
-     * @var mixed
-     */
-    protected $value;
-
     /**
      * Constructor
      * @param string|null $message The string of the error message
      * @param int $code The code of the error
      * @param \Throwable|null $previous the previous exception
-     * @param mixed $value Value that is not a positive
-     * @uses $value
+     * @param mixed $value The value that throwed the exception
      */
     public function __construct($message = null, $code = 0, \Throwable $previous = null, $value = null)
     {
@@ -42,18 +36,6 @@ class NotPositiveException extends Exception
                 $message = sprintf('Value `%s` is not a positive', (string)$value);
             }
         }
-        parent::__construct($message, $code, $previous);
-        $this->value = $value;
-    }
-
-    /**
-     * Gets the value that is not a positive
-     * @return mixed
-     * @since 1.2.11
-     * @uses $value
-     */
-    public function getValue()
-    {
-        return $this->value;
+        parent::__construct($message, $code, $previous, $value);
     }
 }