+++ src/or_fail_functions.php
@@ -28,7 +28,7 @@ if (!function_exists('file_exists_or_fail')) {
      * @param string|null $message The failure message that will be appended to the
      *  generated message
      * @param \Throwable|string $exception The exception class you want to set
-     * @return void
+     * @return string
      * @throws \Tools\Exception\FileNotExistsException
      */
     function file_exists_or_fail($filename, $message = null, $exception = FileNotExistsException::class)
@@ -37,6 +37,8 @@ if (!function_exists('file_exists_or_fail')) {
             $message = sprintf('File or directory `%s` does not exist', rtr($filename));
         }
         is_true_or_fail(is_readable($filename), $message, $exception);
+
+        return $filename;
     }
 }
 
@@ -49,7 +51,7 @@ if (!function_exists('in_array_or_fail')) {
      * @param string $message The failure message that will be appended to
      *  the generated message
      * @param \Throwable|string $exception The exception class you want to set
-     * @return void
+     * @return mixed
      * @since 1.2.6
      */
     function in_array_or_fail($value, $array, $message = null, $exception = NotInArrayException::class)
@@ -58,6 +60,8 @@ if (!function_exists('in_array_or_fail')) {
             $message = sprintf('The value `%s` is not in array', (string)$value);
         }
         is_true_or_fail(in_array($value, $array), $message, $exception);
+
+        return $value;
     }
 }
 
@@ -69,7 +73,7 @@ if (!function_exists('is_dir_or_fail')) {
      * @param string|null $message The failure message that will be appended to the
      *  generated message
      * @param \Throwable|string $exception The exception class you want to set
-     * @return void
+     * @return string
      * @throws \Tools\Exception\NotDirectoryException
      */
     function is_dir_or_fail($filename, $message = null, $exception = NotDirectoryException::class)
@@ -78,6 +82,8 @@ if (!function_exists('is_dir_or_fail')) {
             $message = sprintf('Filename `%s` is not a directory', rtr($filename));
         }
         is_true_or_fail(is_dir($filename), $message, $exception);
+
+        return $filename;
     }
 }
 
@@ -88,7 +94,7 @@ if (!function_exists('is_positive_or_fail')) {
      * @param string $message The failure message that will be appended to the
      *  generated message
      * @param \Throwable|string $exception The exception class you want to set
-     * @return void
+     * @return mixed
      * @since 1.2.5
      * @throws \Tools\Exception\NotPositiveException
      */
@@ -98,6 +104,8 @@ if (!function_exists('is_positive_or_fail')) {
             $message = sprintf('The value `%s` is not a positive', (string)$value);
         }
         is_true_or_fail(is_positive($value), $message, $exception);
+
+        return $value;
     }
 }
 
@@ -109,7 +117,7 @@ if (!function_exists('is_readable_or_fail')) {
      * @param string|null $message The failure message that will be appended to
      *  the generated message
      * @param \Throwable|string $exception The exception class you want to set
-     * @return void
+     * @return string
      * @throws \Tools\Exception\NotReadableException
      */
     function is_readable_or_fail($filename, $message = null, $exception = NotReadableException::class)
@@ -118,6 +126,8 @@ if (!function_exists('is_readable_or_fail')) {
             $message = sprintf('File or directory `%s` is not readable', rtr($filename));
         }
         is_true_or_fail(is_readable($filename), $message, $exception);
+
+        return $filename;
     }
 }
 
@@ -131,14 +141,14 @@ if (!function_exists('is_true_or_fail')) {
      * @param string $message The failure message that will be appended to the
      *  generated message
      * @param \Throwable|string $exception The exception class you want to set
-     * @return void
+     * @return mixed
      * @since 1.1.7
      * @throws \Exception
      */
     function is_true_or_fail($value, $message = 'The value is not equal to `true`', $exception = \ErrorException::class)
     {
         if ($value) {
-            return;
+            return $value;
         }
 
         if ($message instanceof \Exception || (is_string($message) && class_exists($message))) {
@@ -171,7 +181,7 @@ if (!function_exists('is_writable_or_fail')) {
      * @param string|null $message The failure message that will be appended to
      *  the generated message
      * @param \Throwable|string $exception The exception class you want to set
-     * @return void
+     * @return string
      * @throws \Tools\Exception\NotWritableException
      */
     function is_writable_or_fail($filename, $message = null, $exception = NotWritableException::class)
@@ -180,6 +190,8 @@ if (!function_exists('is_writable_or_fail')) {
             $message = sprintf('File or directory `%s` is not writable', rtr($filename));
         }
         is_true_or_fail(is_writable($filename), $message, $exception);
+
+        return $filename;
     }
 }
 
@@ -189,12 +201,12 @@ if (!function_exists('key_exists_or_fail')) {
      *  exception if the key does not exist.
      *
      * If you pass an array of keys, they will all be checked.
-     * @param string|int|array $key Key to check or an array of keys
+     * @param mixed $key Key to check or an array of keys
      * @param array $array An array with keys to check
      * @param \Throwable|string $message The failure message that will be appended to
      *  the generated message
      * @param string $exception The exception class you want to set
-     * @return void
+     * @return mixed
      * @throws \Tools\Exception\KeyNotExistsException
      */
     function key_exists_or_fail($key, array $array, $message = null, $exception = KeyNotExistsException::class)
@@ -206,6 +218,8 @@ if (!function_exists('key_exists_or_fail')) {
             }
             is_true_or_fail($result, $message, $exception);
         }
+
+        return $key;
     }
 }
 
@@ -221,7 +235,7 @@ if (!function_exists('property_exists_or_fail')) {
      * @param string $message The failure message that will be appended to
      *  the generated message
      * @param \Throwable|string $exception The exception class you want to set
-     * @return void
+     * @return string
      * @since 1.1.14
      * @throws \Tools\Exception\PropertyNotExistsException
      */
@@ -234,5 +248,7 @@ if (!function_exists('property_exists_or_fail')) {
             $result = method_exists($object, 'has') ? $object->has($name) : property_exists($object, $name);
             is_true_or_fail($result, $message, $exception);
         }
+
+        return $property;
     }
 }